qmm018: change ExceptionPayRate IntegerPrimitive to DecimalPrimitive, update business on APP

+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/command/rule/employment/averagepay/RegisterAveragePayCommand.java
@@ -1,5 +1,6 @@
 package nts.uk.ctx.pr.core.app.command.rule.employment.averagepay;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Data;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 @NoArgsConstructor
 public class RegisterAveragePayCommand {
 	private int attendDayGettingSet;
-	private int exceptionPayRate;
+	private BigDecimal exceptionPayRate;
 	private int roundDigitSet;
 	private int roundTimingSet;
 	private List<String> selectedSalaryItems;
