+++ nts.uk/ctx/pr/core/nts.uk.ctx.pr.core.app/src/main/java/nts/uk/ctx/pr/core/app/find/rule/employment/averagepay/dto/AveragePayDto.java
@@ -1,5 +1,6 @@
 package nts.uk.ctx.pr.core.app.find.rule.employment.averagepay.dto;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -15,7 +16,7 @@ import nts.uk.ctx.pr.core.app.find.itemmaster.dto.ItemMasterDto;
 public class AveragePayDto {
 	private String companyCode;
 	private int attendDayGettingSet;
-	private int exceptionPayRate;
+	private BigDecimal exceptionPayRate;
 	private int roundDigitSet;
 	private int roundTimingSet;
 	private List<ItemMasterDto> itemsSalary;
